made the daos better and fixed dao-related bugs

+++ pybiz/contrib/redis/dao/redis_types.py
@@ -1,28 +1,28 @@
 from uuid import UUID
 from datetime import datetime
+from threading import local
+from types import GeneratorType
 
+from redis import StrictRedis
 from appyratus.utils import TimeUtils
+
+
+class RedisClient(StrictRedis):
+    pass
 
 
 class RedisObject(object):
     def __init__(self, redis, name):
-        self.redis = redis
-        self.name = name
+        self._redis = redis
+        self._name = name
 
+    @property
+    def redis(self):
+        return self._redis
 
-class Counter(RedisObject):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.reset(value=0)
-
-    def increment(self, value=1):
-        return self.redis.incrby(self.name value)
-
-    def get(self):
-        return self.redis.get(self.name) or 0
-
-    def reset(self, value=0):
-        return self.redis.set(self.name, value)
+    @property
+    def name(self):
+        return self._name
 
 
 class HashSet(RedisObject):
@@ -44,32 +44,54 @@ class HashSet(RedisObject):
     def __len__(self):
         return self.redis.hlen(self.name)
 
-    def increment(self, key, delta=1):
-        return self.redis.hincrby(self.name, key, delta)
+    def increment(self, key, delta=1, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        return redis.hincrby(self.name, key, delta)
 
-    def delete(self, key) -> bool:
-        return bool(seld.redis.hdel(self.name, key))
+    def delete(self, key, pipe=None) -> bool:
+        redis = pipe if pipe is not None else self.redis
+        return bool(pipe.hdel(self.name, key))
 
-    def delete_many(self, keys) -> int:
-        return seld.redis.hdel(self.name, *keys)
+    def delete_many(self, keys, pipe=None) -> int:
+        if isinstance(keys, GeneratorType):
+            keys = tuple(keys)
+        if keys:
+            redis = pipe if pipe is not None else self.redis
+            return redis.hdel(self.name, *keys)
+        return
 
-    def keys(self):
-        return (k.decode() for k in self.redis.hkeys(self.name))
+    def keys(self, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        return (k.decode() for k in redis.hkeys(self.name))
 
-    def values(self):
-        return (v.decode() for v in self.redis.hvals(self.name))
+    def values(self, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        return (v.decode() for v in redis.hvals(self.name))
 
-    def items(self):
-        return zip(self.keys(), self.values())
+    def items(self, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        return zip(self.keys(pipe=pipe), self.values(pipe=pipe))
 
-    def update(self, mapping):
-        self.redis.hmset(self.name, mapping)
+    def update(self, mapping, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        if mapping:
+            redis.hmset(self.name, mapping)
 
-    def get(self, key, default=None):
-        return self.redis.hget(self.name, key) or default
+    def get(self, key, default=None, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        return redis.hget(self.name, key) or default
 
-    def get_many(self, keys):
-        return self.redis.hget(self.name, key) or default
+    def get_many(self, keys, pipe=None):
+        if isinstance(keys, GeneratorType):
+            keys = tuple(keys)
+        if keys:
+            redis = pipe if pipe is not None else self.redis
+            return redis.hmget(self.name, *keys)
+        else:
+            return []
+
+    def get_all(self):
+        return self.redis.hgetall(self.name)
 
 
 class RangeIndex(RedisObject):
@@ -77,13 +99,13 @@ class RangeIndex(RedisObject):
     DELIM = '\0\0'
     DELIM_BYTES = DELIM.encode()
 
-    def upsert(self, _id, value):
+    def upsert(self, _id, value, pipe=None):
         raise NotImplementedError('override in subclass')
 
-    def delete(self, _id):
+    def delete(self, _id, pipe=None):
         raise NotImplementedError('override in subclass')
 
-    def delete_many(self, _ids):
+    def delete_many(self, _ids, pipe=None):
         raise NotImplementedError('override in subclass')
 
     def search(
@@ -93,7 +115,8 @@ class RangeIndex(RedisObject):
         include_lower=True,
         include_upper=False,
         offset=None,
-        limit=None
+        limit=None,
+        pipe=None,
     ):
         raise NotImplementedError('override in subclass')
 
@@ -105,18 +128,24 @@ class NumericIndex(RangeIndex):
         bool: lambda x: int(x),
     }
 
-    def upsert(self, _id, value):
+    def upsert(self, _id, value, pipe=None):
+        redis = pipe if pipe is not None else self.redis
         ser = self.custom_serializers.get(value.__class__)
         value = ser(value) if ser else value
 
-        self.redis.zrem(self.name, _id)
-        self.redis.zadd(self.name, {_id: value})
+        redis.zrem(self.name, _id)
+        redis.zadd(self.name, {_id: value})
 
-    def delete(self, _id):
-        self.redis.zrem(self.name, _id)
+    def delete(self, _id, pipe=None):
+        redis = pipe if pipe is not None else self.redis
+        redis.zrem(self.name, _id)
 
-    def delete_many(self, _ids):
-        self.redis.zrem(self.name, *_ids)
+    def delete_many(self, _ids, pipe=None):
+        if isinstance(_ids, GeneratorType):
+            _ids = tuple(_ids)
+        if _ids:
+            redis = pipe if pipe is not None else self.redis
+            redis.zrem(self.name, *_ids)
 
     def search(
         self,
@@ -125,7 +154,8 @@ class NumericIndex(RangeIndex):
         include_lower=True,
         include_upper=False,
         offset=None,
-        limit=None
+        limit=None,
+        pipe=None
     ):
         if lower is not None:
             lower = '({}'.format(lower) if not include_lower else lower
@@ -140,9 +170,11 @@ class NumericIndex(RangeIndex):
         if limit is not None and offset is None:
             offset = 0
 
+        redis = pipe if pipe is not None else self.redis
+
         return [
             v.split(self.DELIM_BYTES)[-1]
-            for v in self.redis.zrangebyscore(
+            for v in redis.zrangebyscore(
                 self.name, lower, upper, start=offset, num=limit
             )
         ]
@@ -153,26 +185,31 @@ class StringIndex(RangeIndex):
         super().__init__(*args, **kwargs)
         self.lutab = HashSet(self.redis, ':'.join([self.name, 'lutab']))
 
-    def upsert(self, _id, value):
+    def upsert(self, _id, value, pipe=None):
+        redis = pipe if pipe is not None else self.redis
         old_key = self.lutab.get(_id)
         new_key = '{}{}{}'.format(value, self.DELIM, _id)
 
         if old_key is not None:
-            self.redis.zrem(old_key)
+            redis.zrem(old_key)
 
-        self.redis.zadd(self.name, {new_key: 0.0})
+        # TODO: Do this in a pipeline
+        redis.zadd(self.name, {new_key: 0.0})
         self.lutab[_id] = new_key
 
-    def delete(self, _id):
+    def delete(self, _id, pipe=None):
         old_key = self.lutab.get(_id)
-        self.lutab.delete(old_key)
         if old_key is not None:
-            self.redis.zrem(self.name, old_key)
+            redis = pipe if pipe is not None else self.redis
+            self.lutab.delete(old_key)
+            redis.zrem(self.name, old_key)
 
-    def delete_many(self, _ids):
+    def delete_many(self, _ids, pipe=None):
         old_keys = self.lutab.get_many(_ids)
-        self.lutab.delete_many(old_keys)
-        self.redis.zrem(self.name, *old_keys)
+        if old_keys:
+            redis = pipe if pipe is not None else self.redis
+            self.lutab.delete_many(old_keys)
+            redis.zrem(self.name, *old_keys)
 
     def search(
         self,
@@ -181,7 +218,8 @@ class StringIndex(RangeIndex):
         include_lower=False,
         include_upper=True,
         offset=None,
-        limit=None
+        limit=None,
+        pipe=None,
     ):
         if lower is None:
             lower = '-'
@@ -200,7 +238,8 @@ class StringIndex(RangeIndex):
         if limit is not None and offset is None:
             offset = 0
 
-        values = self.redis.zrangebylex(
+        redis = pipe if pipe is not None else self.redis
+        values = redis.zrangebylex(
             self.name, lower, upper, start=offset, num=limit
         )
